requirements.txt
Django~=5.0.6
requests~=2.31.0
python-dotenv==1.0.1
